did implement testcases and added some changes to code for solving some error

+++ serverside/migrations/20230531122432-create-new-table-loginDetails.js
@@ -9,7 +9,7 @@ module.exports = {
         primaryKey: true
       },
       nounce: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(100,20),
       }
     });
   },
